Fix: Tableview crash
Fixed the table view crash on load that was due to a case-sensitive identifier. While this change could have either changed in the xib/storyboard (as it did), it could have also been changed in the  class. The identifier should just match the format of other identifiers in the app. Identifiers may have different rules than assets, so a definition could not be clearly determined based on this single identifier